refactor: rename TelegramUpdate to telegram_update
