+++ src/main/java/com/bot/spider/dtos/telegram_update/FileDTO.java
@@ -1,4 +1,4 @@
-package com.bot.spider.dtos.TelegramUpdate;
+package com.bot.spider.dtos.telegram_update;
 
 public record FileDTO(
         String file_id,
