+++ src/main/java/com/bot/spider/dtos/telegram_update/ForwardOriginDTO.java
@@ -1,4 +1,4 @@
-package com.bot.spider.dtos.TelegramUpdate;
+package com.bot.spider.dtos.telegram_update;
 
 public record ForwardOriginDTO(
         String type,
